requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-core>=0.1.0
openai>=1.3.6
pytest>=7.4.3
requests>=2.31.0
python-dotenv>=1.0.0
allure-pytest>=2.13.2